+++ src/main/java/net/silentchaos512/gear/setup/SgDataComponents.java
@@ -9,16 +9,15 @@ import net.minecraft.util.Unit;
 import net.minecraft.world.item.component.ItemContainerContents;
 import net.neoforged.neoforge.registries.DeferredRegister;
 import net.silentchaos512.gear.SilentGear;
-import net.silentchaos512.gear.api.material.Material;
 import net.silentchaos512.gear.api.part.MaterialGrade;
 import net.silentchaos512.gear.api.part.PartType;
 import net.silentchaos512.gear.api.util.DataResource;
 import net.silentchaos512.gear.core.component.GearConstructionData;
 import net.silentchaos512.gear.core.component.GearPropertiesData;
 import net.silentchaos512.gear.core.component.RepairKitCodecs;
+import net.silentchaos512.gear.core.component.TraitAddedEnchantments;
 import net.silentchaos512.gear.gear.material.MaterialInstance;
 
-import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 import java.util.function.Supplier;
@@ -49,6 +48,12 @@ public class SgDataComponents {
             builder -> builder
                     .persistent(GearPropertiesData.CODEC)
                     .networkSynchronized(GearPropertiesData.STREAM_CODEC)
+    );
+    public static final Supplier<DataComponentType<TraitAddedEnchantments>> TRAIT_ENCHANTMENTS = REGISTRAR.registerComponentType(
+            "trait_enchantments",
+            builder -> builder
+                    .persistent(TraitAddedEnchantments.CODEC)
+                    .networkSynchronized(TraitAddedEnchantments.STREAM_CODEC)
     );
     public static final Supplier<DataComponentType<Boolean>> GEAR_IS_EXAMPLE = REGISTRAR.registerComponentType(
             "is_example",