+++ src/main/java/net/silentchaos512/gear/util/GearData.java
@@ -134,14 +134,13 @@ public final class GearData {
 
         if (oldProperties == null) return;
 
-        // Remove attribute modifiers, so they can be completely redone
+        // Remove data components that need to be completely redone
         gear.remove(DataComponents.ATTRIBUTE_MODIFIERS);
+        gear.remove(SgDataComponents.TRAIT_ENCHANTMENTS);
 
         if (gear.getItem() instanceof GearItem gearItem) {
             gearItem.onRecalculatePre(gear, player, oldProperties, gearConstructionData);
         }
-
-        // TODO: Remove trait-added enchantments
 
         // Let traits do their thing
         for (var trait : oldProperties.getTraits()) {